+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Counted iteration is supported (includung `ExactSizeIterator`) through the `count()`
   "augmentation" methods on the iterators.
 * `derive` feature in `miniconf` crate to control pulling in the derive macros, default enabled
+* Limited depth iteration has been added.
 
 ### Removed
 * [breaking] `TreeKey::iter_*_unchecked()` have been removed. Uncounted iteration is the default.